EW-130 Clean-up of FeathersJS account service (#4012)
* delete feathers account services

* adjust feather auth configuration to work without account service

* remove account endpoints from api validation

* Fix email validation for AdminUser use case

* move trimPassword hook to where it is used

---------
