+++ migrate-etherpads.js
@@ -130,7 +130,7 @@ function chunkArray(myArray, chunk_size) {
  * MAIN
  ****************************************** */
 const run = async (oldPadDomain) => {
-	const app = await appPromise;
+	const app = await appPromise();
 	let searchRegex = new RegExp(`https://${oldPadDomain.replace(/\./g, '\\.')}.*`);
 
 	const lessonsService = app.service('/lessons');
