+++ src/services/authentication/configuration.js
@@ -4,8 +4,10 @@ const { authenticationSecret, audience } = require('./logic/index');
 module.exports = {
 	audience,
 	authConfig: {
-		entity: 'account',
-		service: 'accountModel',
+		entity: 'account', // name of the found user in the request context
+		// entityId and service are never queried, but need to be provided otherwise the server doesn't start
+		entityId: 'noId',
+		service: 'emptyService', // This service is registered in 'index.js'
 		secret: authenticationSecret,
 		authStrategies: ['jwt', 'tsp', 'api-key'],
 		jwtOptions: {
